Adjust fracturing to be fixed joint
